+++ Box.V2/BoxClient.cs
@@ -135,6 +135,7 @@ namespace Box.V2
             CollaborationWhitelistManager = new BoxCollaborationWhitelistManager(Config, _service, _converter, Auth, _asUser, _suppressNotifications);
             TermsOfServiceManager = new BoxTermsOfServiceManager(Config, _service, _converter, Auth, _asUser, _suppressNotifications);
             MetadataCascadePolicyManager = new BoxMetadataCascadePolicyManager(Config, _service, _converter, Auth, _asUser, _suppressNotifications);
+            StoragePoliciesManager = new BoxStoragePoliciesManager(Config, _service, _converter, Auth, _asUser, _suppressNotifications);
 
             // Init Resource Plugins Manager
             ResourcePlugins = new BoxResourcePlugins();
@@ -271,5 +272,9 @@ namespace Box.V2
         /// The manager that represents the metadata cascade policy endpoint
         /// </summary>
         public BoxMetadataCascadePolicyManager MetadataCascadePolicyManager { get; private set; }
+
+        /// The manager that represents the storage policies endpoint
+        /// </summary>
+        public BoxStoragePoliciesManager StoragePoliciesManager { get; private set; }
     }
 }