+++ Box.V2/Config/Constants.cs
@@ -50,6 +50,9 @@ namespace Box.V2.Config
         public const string TermsOfServicesString = @"terms_of_services/";
         public const string TermsOfServiceUserStatusesString = @"terms_of_service_user_statuses/";
         public const string MetadataCascadePoliciesString = @"metadata_cascade_policies/";
+        public const string StoragePoliciesString = @"storage_policies/";
+        public const string StoragePolicyAssignmentsString = @"storage_policy_assignments/";
+        public const string StoragePolicyAssignmentsForTargetString = @"storage_policy_assignments";
 
         /// <summary>
         /// The shared items constant
@@ -101,6 +104,8 @@ namespace Box.V2.Config
         public const string LegalHoldPoliciesEndpointString = BoxApiUriString + LegalHoldPoliciesString;
         public const string LegalHoldPolicyAssignmentsEndpointString = BoxApiUriString + LegalHoldPolicyAssignmentsString;
         public const string MetadataCascadePolicyEndpointString = BoxApiUriString + MetadataCascadePoliciesString;
+        public const string StoragePoliciesEndpointString = BoxApiUriString + StoragePoliciesString;
+        public const string StoragePolicyAssignmentsEndpointString = BoxApiUriString + StoragePolicyAssignmentsString;
 
         /*** Endpoint Paths ***/
         public const string ItemsPathString = @"{0}/items";
@@ -171,6 +176,9 @@ namespace Box.V2.Config
         public const string TypeTermsOfService = "terms_of_service";
         public const string TypeTermsOfServiceUserStatuses = "terms_of_service_user_status";
         public const string TypeMetadataCascadePolicy = "metadata_cascade_policy";
+        public const string TypeStoragePolicy = "storage_policy";
+        public const string TypeStoragePolicyAssignment = "storage_policy_assignment";
+        public const string TypeApplication = "application";
 
         /*** File Preview ***/
         public const int DefaultRetryDelay = 1000; // milliseconds
@@ -497,7 +505,7 @@ namespace Box.V2.Config
             /// Added device association.
             /// </summary>
             public const string AddDeviceAssociation = "ADD_DEVICE_ASSOCIATION";
-            
+
             /// <summary>
             /// Edit the permissions on a folder. 
             /// </summary>
@@ -852,8 +860,13 @@ namespace Box.V2.Config
             /// An OAuth2 access token was generated for a user.
             /// </summary>
             public const string UserAuthenticateOAuth2AccessTokenCreate = "USER_AUTHENTICATE_OAUTH2_ACCESS_TOKEN_CREATE";
+
+            /// <summary>
+            /// A content was accessed by a user.
+            /// </summary>
+            public const string ContentAccess = "CONTENT_ACCESS";
         }
-        
+
         /*** required direction for collaboration whitelists endpoint. Can only be one of value: inbound, outbound, or both ***/
         public static class WhitelistDirections
         {